requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
langchain==0.2.16
langchain-google-genai==1.0.10
python-multipart==0.0.12
python-dotenv==1.0.1
